Remove ViewPager and Slider in favor of exporting createNativeWrapper (#506)
Fixes #482 

In RN 0.59 `ViewPager` and `Slider` are going to be dropped and therefore they call for being dropped from RNGH core. However, since there're use cases of these components, decided not to drop support for them at all, but to expose function for creating generic components with pure native ones.
