+++ Example/android/app/src/main/java/com/swmansion/gesturehandler/react/example/MainApplication.java
@@ -3,6 +3,8 @@ package com.swmansion.gesturehandler.react.example;
 import android.app.Application;
 
 import com.facebook.react.ReactApplication;
+import com.reactnativecommunity.viewpager.RNCViewPagerPackage;
+import com.reactnativecommunity.slider.ReactSliderPackage;
 import com.oblador.vectoricons.VectorIconsPackage;
 import com.facebook.react.ReactNativeHost;
 import com.facebook.react.ReactPackage;
@@ -25,6 +27,8 @@ public class MainApplication extends Application implements ReactApplication {
     protected List<ReactPackage> getPackages() {
       return Arrays.<ReactPackage>asList(
           new MainReactPackage(),
+            new RNCViewPagerPackage(),
+            new ReactSliderPackage(),
             new VectorIconsPackage(),
           new RNGestureHandlerPackage()
       );
