+++ docs/about-handlers.md
@@ -77,11 +77,9 @@ class Multitap extends Component {
 Gesture handler library exposes a set of components normally available in React Native that are wrapped in [`NativeViewGestureHandler`](handler-native.md).
 Here is a list of exposed components:
  - `ScrollView`
- - `Slider`
  - `Switch`
  - `TextInput`
  - `ToolbarAndroid` (**Android only**)
- - `ViewPagerAndroid` (**Android only**)
  - `DrawerLayoutAndroid` (**Android only**)
  
 If you want to use other handlers or [buttons](component-buttons.md) nested in a `ScrollView` or you want to use [`waitFor`](handler-common.md#waitfor) property to define interaction between a handler and `ScrollView`
